+++ esi/tests/test_models.py
@@ -1,7 +1,6 @@
 """unit tests for esi"""
 
 from datetime import timedelta
-import logging
 from unittest.mock import patch, Mock
 
 from django.core.exceptions import ImproperlyConfigured
@@ -13,42 +12,126 @@ from oauthlib.oauth2.rfc6749.errors import InvalidGrantError, \
     MissingTokenError, InvalidClientError, InvalidTokenError, \
     InvalidClientIdError
 
-from . import _generate_token, _store_as_Token, _set_logger
-from ..errors import TokenInvalidError, NotRefreshableTokenError, \
-    TokenExpiredError, IncompleteResponseError
-from ..models import Scope, Token, CallbackRedirect
+from esi.errors import TokenInvalidError, NotRefreshableTokenError, \
+    TokenExpiredError, IncompleteResponseError, TokenError
+from esi.models import Token
 
-logger = _set_logger(logging.getLogger('esi.models'), __file__)
+from . import _generate_token, _store_as_Token
+from .factories_2 import CallbackRedirectFactory, ScopeFactory, TokenFactory
+
+
+MODULE_PATH = "esi.models"
 
 
 class TestScope(TestCase):
-    """tests for model Scope"""
+    def test_should_return_str(self):
+        # given
+        obj = ScopeFactory.build(name="dummy_scope")
+        # when/then
+        self.assertEqual(str(obj), "dummy_scope")
 
-    def test_str(self):
-        x = Scope(name='dummy_scope')
-        self.assertEqual(
-            "dummy_scope",
-            str(x)
-        )
+    def test_should_return_friendly_name_1(self):
+        # given
+        obj = ScopeFactory.build(name="dummy_scope")
+        # when/then
+        self.assertEqual(obj.friendly_name, "dummy_scope")
 
-    def test_friendly_name_fail(self):
-        x = Scope(name='dummy_scope')
-        self.assertEqual(
-            "dummy_scope",
-            x.friendly_name
-        )
+    def test_should_return_friendly_name_2(self):
+        # given
+        obj = ScopeFactory.build(name="test.dummy_scope.test")
+        # when/then
+        self.assertEqual(obj.friendly_name, "dummy scope")
 
-    def test_friendly_name_pass(self):
-        x = Scope(name='test.dummy_scope.test')
-        self.assertEqual(
-            "dummy scope",
-            x.friendly_name
+
+class TestToken2(TestCase):
+    def test_should_return_str(self):
+        # given
+        scope = ScopeFactory(name="dummy_scope")
+        token = TokenFactory(character_name="Bruce Wayne", scopes=[scope])
+        # when/then
+        self.assertEqual(str(token), "Bruce Wayne - dummy_scope")
+
+    def test_should_return_str_not_created(self):
+        # given
+        token = TokenFactory.build(character_name="Bruce Wayne")
+        # when/then
+        self.assertEqual(str(token), "Bruce Wayne - ")
+
+    def test_should_return_repr(self):
+        # given
+        scope = ScopeFactory(name="dummy_scope")
+        token = TokenFactory(
+            id=99, character_id=1001, character_name="Bruce Wayne", scopes=[scope]
         )
+        # when/then
+        self.assertEqual(repr(token), "<Token(id=99): 1001, Bruce Wayne>")
+
+    def test_should_find_token(self):
+        # given
+        scope = ScopeFactory(name="dummy_scope")
+        token = TokenFactory(character_id=1001, scopes=[scope])
+        TokenFactory()
+        # when
+        result = Token.get_token(1001, ['dummy_scope'])
+        # then
+        self.assertEqual(token, result)
+
+    def test_should_return_false_when_token_not_found(self):
+        # given
+        scope = ScopeFactory(name="dummy_scope")
+        TokenFactory(character_id=1001, scopes=[scope])
+        # when
+        result = Token.get_token(1001, ['unknown'])
+        # then
+        self.assertFalse(result)
+
+    def test_should_be_able_to_refresh(self):
+        # given
+        token = TokenFactory.build()
+        # when/then
+        self.assertTrue(token.can_refresh)
+
+    def test_should_not_be_able_to_refresh(self):
+        # given
+        token = TokenFactory.build(refresh_token=None)
+        # when/then
+        self.assertFalse(token.can_refresh)
+
+
+@patch(MODULE_PATH + ".Token.refresh", spec=True)
+class TestTokenRefreshOrDelete(TestCase):
+    def test_should_refresh_token(self, mock_token_refresh):
+        # given
+        token = TokenFactory()
+        # when
+        token.refresh_or_delete()
+        # then
+        self.assertTrue(mock_token_refresh.called)
+        self.assertTrue(Token.objects.filter(pk=token.pk).exists())
+
+    def test_should_delete_token_with_errors(self, mock_token_refresh):
+        # given
+        mock_token_refresh.side_effect = TokenError
+        token = TokenFactory()
+        # when
+        token.refresh_or_delete()
+        # then
+        self.assertTrue(mock_token_refresh.called)
+        self.assertFalse(Token.objects.filter(pk=token.pk).exists())
+
+    def test_should_not_delete_token_with_incomplete_response_error(self, mock_token_refresh):
+        # given
+        mock_token_refresh.side_effect = IncompleteResponseError
+        token = TokenFactory()
+        # when
+        with self.assertRaises(IncompleteResponseError):
+            token.refresh_or_delete()
+        # then
+        self.assertTrue(mock_token_refresh.called)
+        self.assertTrue(Token.objects.filter(pk=token.pk).exists())
 
 
 class TestToken(TestCase):
-    """tests for model Token"""
-
     def setUp(self):
 
         character_id = 1000
@@ -68,45 +151,18 @@ class TestToken(TestCase):
             self.user
         )
 
-    def test_str(self):
-        self.assertEqual(
-            "Bruce Wayne - esi-universe.read_structures.v1",
-            str(self.token)
-        )
-
-    def test_repr(self):
-        self.assertEqual(
-            '<Token(id={}): 1000, Bruce Wayne>'.format(
-                self.token.pk
-            ),
-            repr(self.token)
-        )
-
-    def test_get_token(self):
-        t = Token.get_token(1000, ['esi-universe.read_structures.v1'])
-        self.assertEqual(t, self.token)
-
-    def test_get_token_fail(self):
-        t = Token.get_token(1000, ['esi-universe.read_structures.v9000'])
-        self.assertEqual(t, False)
-
-    def test_need_refresh_token_for_refresh(self):
-        self.assertTrue(self.token.can_refresh)
-        self.token.refresh_token = None
-        self.assertFalse(self.token.can_refresh)
-
-    @patch('esi.models.app_settings.ESI_TOKEN_VALID_DURATION', 120)
+    @patch(MODULE_PATH + '.app_settings.ESI_TOKEN_VALID_DURATION', 120)
     def test_expires(self):
         self.assertEqual(
             self.token.created + timedelta(seconds=120),
             self.token.expires
         )
 
-    @patch('esi.models.app_settings.ESI_TOKEN_VALID_DURATION', 120)
+    @patch(MODULE_PATH + '.app_settings.ESI_TOKEN_VALID_DURATION', 120)
     def test_not_expired(self):
         self.assertFalse(self.token.expired)
 
-    @patch('esi.models.app_settings.ESI_TOKEN_VALID_DURATION', 120)
+    @patch(MODULE_PATH + '.app_settings.ESI_TOKEN_VALID_DURATION', 120)
     def test_has_expired(self):
         self.token.created -= timedelta(121)
         self.assertTrue(self.token.expired)
@@ -130,8 +186,8 @@ class TestToken(TestCase):
             self.token.created, timezone.now() - timedelta(seconds=60)
         )
 
-    @patch('esi.models.HTTPBasicAuth', autospec=True)
-    @patch('esi.models.OAuth2Session', autospec=True)
+    @patch(MODULE_PATH + '.HTTPBasicAuth', autospec=True)
+    @patch(MODULE_PATH + '.OAuth2Session', autospec=True)
     def test_refresh_normal_2(self, mock_OAuth2Session, mock_HTTPBasicAuth):
         mock_session = Mock()
         mock_session.refresh_token.return_value = {
@@ -157,9 +213,9 @@ class TestToken(TestCase):
         self.assertFalse(self.token.expired)
         self.assertEqual(self.token.valid_access_token(), 'access_token')
 
-    @patch('esi.models.HTTPBasicAuth', autospec=True)
-    @patch('esi.models.OAuth2Session', autospec=True)
-    @patch('esi.models.app_settings.ESI_TOKEN_VALID_DURATION', 120)
+    @patch(MODULE_PATH + '.HTTPBasicAuth', autospec=True)
+    @patch(MODULE_PATH + '.OAuth2Session', autospec=True)
+    @patch(MODULE_PATH + '.app_settings.ESI_TOKEN_VALID_DURATION', 120)
     def test_valid_access_token_refresh(self, mock_OAuth2Session, mock_HTTPBasicAuth):
         mock_session = Mock()
         mock_session.refresh_token.return_value = {
@@ -173,9 +229,9 @@ class TestToken(TestCase):
         self.assertEqual(
             self.token.valid_access_token(), 'access_token_new')
 
-    @patch('esi.models.HTTPBasicAuth', autospec=True)
-    @patch('esi.models.OAuth2Session', autospec=True)
-    @patch('esi.models.app_settings.ESI_TOKEN_VALID_DURATION', 120)
+    @patch(MODULE_PATH + '.HTTPBasicAuth', autospec=True)
+    @patch(MODULE_PATH + '.OAuth2Session', autospec=True)
+    @patch(MODULE_PATH + '.app_settings.ESI_TOKEN_VALID_DURATION', 120)
     def test_valid_access_token_cant_refresh(
         self, mock_OAuth2Session, mock_HTTPBasicAuth
     ):
@@ -220,7 +276,7 @@ class TestToken(TestCase):
         with self.assertRaises(ImproperlyConfigured):
             self.token.refresh(mock_session, mock_auth)
 
-    @patch('esi.models.esi_client_factory', autospec=True)
+    @patch(MODULE_PATH + '.esi_client_factory', autospec=True)
     def test_get_esi_client(self, mock_esi_client):
         mock_esi_client.return_value = "Johnny"
         x = self.token.get_esi_client()
@@ -242,7 +298,7 @@ class TestToken(TestCase):
         data = self.token.get_token_data(access_token='access_token_2')
         self.assertEqual(data['name'], "Bruce Wayne")
     """
-    @patch('esi.models.Token.get_token_data')
+    @patch(MODULE_PATH + '.Token.get_token_data')
     def test_update_token_data_normal_1(self, mock_get_token_data):
         mock_get_token_data.return_value = _generate_token(99, 'Bruce Wayne')
         self.token.update_token_data()
@@ -264,9 +320,9 @@ class TestToken(TestCase):
             'character'
         )
 
-    @patch('esi.models.HTTPBasicAuth', autospec=True)
-    @patch('esi.models.OAuth2Session', autospec=True)
-    @patch('esi.models.Token.get_token_data')
+    @patch(MODULE_PATH + '.HTTPBasicAuth', autospec=True)
+    @patch(MODULE_PATH + '.OAuth2Session', autospec=True)
+    @patch(MODULE_PATH + '.Token.get_token_data')
     def test_update_token_data_normal_2(
         self,
         mock_get_token_data,
@@ -295,7 +351,7 @@ class TestToken(TestCase):
             99
         )
 
-    @patch('esi.models.Token.get_token_data')
+    @patch(MODULE_PATH + '.Token.get_token_data')
     def test_update_token_data_normal_3(self, mock_get_token_data):
         mock_get_token_data.return_value = {
             'character_id': 99,
@@ -321,7 +377,7 @@ class TestToken(TestCase):
             'Character'
         )
 
-    @patch('esi.models.Token.get_token_data', auto_spec=True, unsafe=True)
+    @patch(MODULE_PATH + '.Token.get_token_data', auto_spec=True, unsafe=True)
     def test_update_token_data_error(
         self,
         mock_get_token_data
@@ -333,25 +389,15 @@ class TestToken(TestCase):
 
 
 class TestCallbackRedirect(TestCase):
+    def test_should_return_str(self):
+        # given
+        cb = CallbackRedirectFactory.build(session_key="abc", url="/green/alpha")
+        # when/then
+        self.assertEqual("abc: /green/alpha", str(cb))
 
-    @classmethod
-    def setUpClass(cls):
-        super().setUpClass()
-        cls.redirect_url = 'https://www.example.com/%s/' % ('x' * (2048 - 25))
-        cls.cb = CallbackRedirect.objects.create(
-            session_key='abc',
-            state='xyz',
-            url=cls.redirect_url
-        )
-        cls.cb.refresh_from_db()
-
-    def test_str(self):
-        self.assertEqual('abc: %s' % self.redirect_url, str(self.cb))
-
-    def test_repr(self):
-        self.assertEqual(
-            '<CallbackRedirect(pk={}): abc to {}>'.format(
-                self.cb.pk, self.redirect_url
-            ),
-            repr(self.cb)
-        )
+    def test_should_return_repr(self):
+        # given
+        cb = CallbackRedirectFactory(session_key="abc", url="/green/alpha")
+        # when/then
+        expected = f"<CallbackRedirect(pk={cb.pk}): abc to /green/alpha>"
+        self.assertEqual(expected, repr(cb))