+++ tests/celery.py
@@ -1,0 +1,17 @@
+import os
+from celery import Celery
+
+# set the default Django settings module for the 'celery' program.
+os.environ.setdefault("DJANGO_SETTINGS_MODULE", "tests.settings")
+
+from django.conf import settings  # noqa
+
+app = Celery("mysite", broker_url="memory://localhost/")
+app.conf.update(CELERY_ALWAYS_EAGER=True, CELERY_EAGER_PROPAGATES_EXCEPTIONS=True)
+
+# Using a string here means the worker don't have to serialize
+# the configuration object to child processes.
+app.config_from_object("django.conf:settings")
+
+# Load task modules from all registered Django app configs.
+app.autodiscover_tasks(lambda: settings.INSTALLED_APPS)